requirements.txt
Flask
Flask-Bcrypt
Flask-Login
gunicorn
python-dotenv
qrcode
Pillow
Flask-SQLAlchemy
psycopg2-binary